Interpolation experiment & ellipse size
I realised that I had a parameter still for the ellipse size, but this was outdated because I decided to keep the ellipse size consistent in the end. Therefore, I removed this parameter, so I now have one less parameter which is great! I also experimented with the interpolation feature to see what I could come up with.