Lots of changes
I was working on all the details today ready for hand-in. So there was a lot of trial and error with finicky details. I did many iterations for the default letterform, eventually landing on one. I tidied up my code, making sure to remove any unnecessary code that has been blanked out, and ordering my readme document chronologically to when I did stuff.